requirements.txt
opencv-python
ultralytics
numpy
torch
ncnn
RPi.GPIO
gpiozero
